+++ src/queries.jl
@@ -110,6 +110,8 @@ julia> unitknot[Lift((a=(x=1,y=2),)) >> It.a.x]
 """
 const It = Navigation(())
 
+translate(mod::Module, ::Val{:it}) = It
+
 
 #
 # Querying a DataKnot.
@@ -128,7 +130,7 @@ query(db, F; kws...) =
 
 function query(db::DataKnot, F::AbstractQuery, params::Vector{Pair{Symbol,DataKnot}}=Pair{Symbol,DataKnot}[])
     db = pack(db, params)
-    q = assemble(F, shape(db))
+    q = assemble(shape(db), F)
     db′ = q(db)
     return db′
 end
@@ -144,13 +146,121 @@ function pack(db::DataKnot, params::Vector{Pair{Symbol,DataKnot}})
 end
 
 assemble(db::DataKnot, F::AbstractQuery, params::Vector{Pair{Symbol,DataKnot}}=Pair{Symbol,DataKnot}[]) =
-    assemble(F, shape(pack(db, params)))
+    assemble(shape(pack(db, params)), F)
 
-function  assemble(F::AbstractQuery, src::AbstractShape)
-    env = Environment()
-    q = uncover(assemble(F, env, cover(src)))
-    return optimize(q)
+assemble(src::AbstractShape, F::AbstractQuery) =
+    optimize(uncover(assemble(nothing, cover(src), F)))
+
+assemble(::Nothing, p::Pipeline, F) =
+    assemble(Environment(), p, F)
+
+#
+# External syntax.
+#
+"""
+     @query expr
+
+Creates a query object from a specialized path-like notation:
+
+* bare identifiers are translated to navigation with `Get`;
+* query combinators, such as `Count(X)`, use lower-case names;
+* the period (`.`) is used for query composition (`>>`);
+* aggregate queries, such as `Count`, require parentheses;
+* records can be constructed using curly brackets, `{}`; and
+* functions and operators are lifted automatically.
+
+```jldoctest
+julia> @query 2x+1
+Lift(+, (Lift(*, (Lift(2), Get(:x))), Lift(1)))
+```
+"""
+macro query(ex)
+    return :( translate($__module__, $(Expr(:quote, ex)) ) )
 end
+
+"""
+     @query dataset expr param=...
+
+Applies the query to a dataset with a given set of parameters.
+
+```jldoctest
+julia> @query unitknot 2x+1 x=1
+│ It │
+┼────┼
+│  3 │
+```
+"""
+macro query(db, exs...)
+    exs = map(exs) do ex
+        if Meta.isexpr(ex, :(=), 2)
+            esc(Expr(:kw, ex.args...))
+        else
+            :( Each(translate($__module__, $(Expr(:quote, ex)) )) )
+        end
+    end
+    return quote
+        query($(esc(db)), $(exs...))
+    end
+end
+
+function translate(mod::Module, ex::Expr)::AbstractQuery
+    head = ex.head
+    args = ex.args
+    if head === :block
+        return Compose(translate.(Ref(mod), filter(arg -> !(arg isa LineNumberNode), args))...)
+    elseif head === :. && length(args) == 2 && Meta.isexpr(args[2], :tuple, 1)
+        return Compose(translate(mod, args[1]), translate(mod, args[2].args[1]))
+    elseif head === :.
+        return Compose(translate.(Ref(mod), args)...)
+    elseif head === :braces
+        return Record(translate.(Ref(mod), args)...)
+    elseif head === :quote && length(args) == 1 && Meta.isexpr(args[1], :braces)
+        return Record(translate.(Ref(mod), args[1].args)...)
+    elseif head === :curly && length(args) >= 1
+        return Compose(translate(mod, args[1]), Record(translate.(Ref(mod), args[2:end])...))
+    elseif head === :call && length(args) >= 1
+        call = args[1]
+        if call === :(=>) && length(args) == 3 && args[2] isa Symbol
+            return Compose(translate(mod, args[3]), Label(args[2]))
+        elseif call isa Symbol
+            return translate(mod, Val(call), (args[2:end]...,))
+        elseif call isa QuoteNode && call.value isa Symbol
+            return translate(mod, Val(call.value), (args[2:end]...,))
+        elseif Meta.isexpr(call, :.) && !isempty(call.args)
+            return Compose(translate.(Ref(mod), call.args[1:end-1])...,
+                           translate(mod, Expr(:call, call.args[end], args[2:end]...)))
+        end
+    end
+    error("invalid query expression: $(repr(ex))")
+end
+
+translate(mod::Module, sym::Symbol) =
+    translate(mod, Val(sym))
+
+translate(::Module, ::Val{:nothing}) =
+    Lift(nothing)
+
+translate(::Module, ::Val{:missing}) =
+    Lift(missing)
+
+translate(mod::Module, qn::QuoteNode) =
+    translate(mod, qn.value)
+
+translate(mod::Module, @nospecialize(v::Val{N})) where {N} =
+    Get(N)
+
+function translate(mod::Module, @nospecialize(v::Val{N}), args::Tuple) where {N}
+    fn = getfield(mod, N)
+    oty = Core.Compiler.return_type(fn, Tuple{map(arg -> Query, args)...})
+    if oty != Union{} && oty <: Union{AbstractQuery,Pair{Symbol,<:AbstractQuery}}
+        return Lift(fn(translate.(Ref(mod), args)...))
+    else
+        return Lift(fn, translate.(Ref(mod), args))
+    end
+end
+
+translate(mod::Module, val) =
+    Lift(val)
 
 
 #
@@ -165,13 +275,13 @@ Query compilation state.
 mutable struct Environment
 end
 
-assemble(F, env::Environment, p::Pipeline)::Pipeline =
-    assemble(Lift(F), env, p)
+assemble(env::Environment, p::Pipeline, F)::Pipeline =
+    assemble(env, p, Lift(F))
 
-assemble(F::Query, env::Environment, p::Pipeline)::Pipeline =
+assemble(env::Environment, p::Pipeline, F::Query)::Pipeline =
     F.op(env, p, F.args...)
 
-function assemble(nav::Navigation, env::Environment, p::Pipeline)::Pipeline
+function assemble(env::Environment, p::Pipeline, nav::Navigation)::Pipeline
     for name in getfield(nav, :__path)
         p = Get(env, p, name)
     end
@@ -430,15 +540,15 @@ end
 >>(X::Union{DataKnot,AbstractQuery,Pair{Symbol,<:Union{DataKnot,AbstractQuery}}}, Xs...) =
     Compose(X, Xs...)
 
-Compose(X, Xs...) =
-    Query(Compose, X, Xs...)
+Compose(Xs...) =
+    Query(Compose, Xs...)
 
 quoteof(::typeof(Compose), args::Vector{Any}) =
     quoteof(>>, args)
 
 function Compose(env::Environment, p::Pipeline, Xs...)
     for X in Xs
-        p = assemble(X, env, p)
+        p = assemble(env, p, X)
     end
     p
 end
@@ -514,9 +624,12 @@ Record(Xs...) =
     Query(Record, Xs...)
 
 function Record(env::Environment, p::Pipeline, Xs...)
-    xs = assemble.(collect(AbstractQuery, Xs), Ref(env), Ref(target_pipe(p)))
+    xs = assemble.(Ref(env), Ref(target_pipe(p)), collect(AbstractQuery, Xs))
     assemble_record(p, xs)
 end
+
+translate(mod::Module, ::Val{:record}, args::Tuple) =
+    Record(translate.(Ref(mod), args)...)
 
 
 #
@@ -610,40 +723,28 @@ In the combinator form, `Collect(X₁, X₂ … Xₙ)` adds fields `X₁`,
 `X₂` … `Xₙ` to the input record.
 
 ```jldoctest
-julia> unitknot[Lift(1:3) >> Record(It) >> Collect(It.A .* It.A)]
-  │ #A  #B │
-──┼────────┼
-1 │  1   1 │
-2 │  2   4 │
-3 │  3   9 │
-```
-
-Input `Tuple` and `NamedTuple` are promoted to `Record`.
-
-```jldoctest
-julia> unitknot[Lift((a=2,)) >> Collect(:b=>It.a .* It.a)]
-│ a  b │
+julia> unitknot[Record(:x => 1) >> Collect(:y => 2 .* It.x)]
+│ x  y │
 ┼──────┼
-│ 2  4 │
+│ 1  2 │
 ```
 
-Since the `unitknot` is an empty tuple, it is treated as a record.
+If a field already exists, it is replaced.
 
 ```jldoctest
-julia> unitknot[Collect(:a=>2, :b=>It.a .* It.a)]
-│ a  b │
-┼──────┼
-│ 2  4 │
-```
-
-If a label is assigned the value `nothing` then it is removed.
-
-```jldoctest
-julia> unitknot[Lift((a=2,)) >>
-                Collect(:b=>It.a .* It.a, :a => nothing)]
-│ b │
+julia> unitknot[Record(:x => 1) >> Collect(:x => 2 .* It.x)]
+│ x │
 ┼───┼
-│ 4 │
+│ 2 │
+```
+
+To remove a field, assign it the value `nothing`.
+
+```jldoctest
+julia> unitknot[Record(:x => 1) >> Collect(:y => 2 .* It.x, :x => nothing)]
+│ y │
+┼───┼
+│ 2 │
 ```
 
 ---
@@ -653,12 +754,10 @@ julia> unitknot[Lift((a=2,)) >>
 In the query form, `Collect` appends a field to the origin record.
 
 ```jldoctest
-julia> X = Lift('a':'c');
-
-julia> unitknot[X >> Collect]
-│ #A      │
-┼─────────┼
-│ a; b; c │
+julia> unitknot[Lift(1) >> Label(:x) >> Collect]
+│ x │
+┼───┼
+│ 1 │
 ```
 """
 Collect(X, Ys...) =
@@ -671,9 +770,15 @@ Collect(env::Environment, p::Pipeline, X, Ys...) =
     Collect(env, Collect(env, p, X), Ys...)
 
 function Collect(env::Environment, p::Pipeline, X)
-    x = assemble(X, env, target_pipe(p))
+    x = assemble(env, target_pipe(p), X)
     assemble_collect(p, x)
 end
+
+translate(mod::Module, ::Val{:collect}, args::Tuple) =
+    Collect(translate.(Ref(mod), args)...)
+
+translate(mod::Module, ::Val{:collect}, ::Tuple{}) =
+    Then(Collect)
 
 
 #
@@ -811,7 +916,7 @@ Lift(env::Environment, p::Pipeline, elts::AbstractVector, card::Union{Cardinalit
     Lift(env, p, DataKnot(Any, elts, card))
 
 function Lift(env::Environment, p::Pipeline, f, Xs::Tuple)
-    xs = assemble.(collect(AbstractQuery, Xs), Ref(env), Ref(target_pipe(p)))
+    xs = assemble.(Ref(env), Ref(target_pipe(p)), collect(AbstractQuery, Xs))
     assemble_lift(p, f, xs)
 end
 
@@ -891,7 +996,10 @@ julia> unitknot[Lift(1:3) >> X]
 Each(X) = Query(Each, X)
 
 Each(env::Environment, p::Pipeline, X) =
-    compose(p, assemble(X, env, target_pipe(p)))
+    compose(p, assemble(env, target_pipe(p), X))
+
+translate(mod::Module, ::Val{:each}, args::Tuple{Any}) =
+    Each(translate(mod, args[1]))
 
 
 #
@@ -927,6 +1035,9 @@ Label(env::Environment, p::Pipeline, lbl::Symbol) =
 
 Lift(p::Pair{Symbol}) =
     Compose(p.second, Label(p.first))
+
+translate(mod::Module, ::Val{:label}, args::Tuple{Symbol}) =
+    Label(args[1])
 
 
 #
@@ -974,10 +1085,10 @@ Tag(F::Union{Function,DataType}, args::Tuple, X) =
     Tag(nameof(F), args, X)
 
 Tag(env::Environment, p::Pipeline, name, X) =
-    assemble(X, env, p)
+    assemble(env, p, X)
 
 Tag(env::Environment, p::Pipeline, name, args, X) =
-    assemble(X, env, p)
+    assemble(env, p, X)
 
 quoteof(::typeof(Tag), args::Vector{Any}) =
     quoteof(Tag, args...)
@@ -1170,9 +1281,12 @@ Keep(env::Environment, p::Pipeline, P, Qs...) =
     Keep(env, Keep(env, p, P), Qs...)
 
 function Keep(env::Environment, p::Pipeline, P)
-    q = assemble(P, env, target_pipe(p))
+    q = assemble(env, target_pipe(p), P)
     assemble_keep(p, q)
 end
+
+translate(mod::Module, ::Val{:keep}, args::Tuple{Any,Vararg{Any}}) =
+    Keep(translate.(Ref(mod), args)...)
 
 
 #
@@ -1204,9 +1318,12 @@ Given(env::Environment, p::Pipeline, Xs...) =
     Given(env, p, Keep(Xs[1:end-1]...) >> Each(Xs[end]))
 
 function Given(env::Environment, p::Pipeline, X)
-    q = assemble(X, env, target_pipe(p))
+    q = assemble(env, target_pipe(p), X)
     assemble_given(p, q)
 end
+
+translate(mod::Module, ::Val{:given}, args::Tuple{Any,Vararg{Any}}) =
+    Given(translate.(Ref(mod), args)...)
 
 
 #
@@ -1226,7 +1343,7 @@ Then(ctor, args::Tuple) =
     Query(Then, ctor, args)
 
 Then(env::Environment, p::Pipeline, ctor, args::Tuple=()) =
-    assemble(ctor(Then(p), args...), env, source_pipe(p))
+    assemble(env, source_pipe(p), ctor(Then(p), args...))
 
 
 #
@@ -1291,7 +1408,7 @@ Lift(::typeof(Count)) =
     Then(Count)
 
 function Count(env::Environment, p::Pipeline, X)
-    x = assemble(X, env, target_pipe(p))
+    x = assemble(env, target_pipe(p), X)
     compose(p, assemble_count(x))
 end
 
@@ -1429,7 +1546,7 @@ Lift(::typeof(Min)) =
     Then(Min)
 
 function Sum(env::Environment, p::Pipeline, X)
-    x = assemble(X, env, target_pipe(p))
+    x = assemble(env, target_pipe(p), X)
     assemble_lift(p, sum, Pipeline[x])
 end
 
@@ -1437,7 +1554,7 @@ maximum_missing(v) =
     !isempty(v) ? maximum(v) : missing
 
 function Max(env::Environment, p::Pipeline, X)
-    x = assemble(X, env, target_pipe(p))
+    x = assemble(env, target_pipe(p), X)
     card = cardinality(target(x))
     optional = fits(x0to1, card)
     assemble_lift(p, optional ? maximum_missing : maximum, Pipeline[x])
@@ -1447,11 +1564,35 @@ minimum_missing(v) =
     !isempty(v) ? minimum(v) : missing
 
 function Min(env::Environment, p::Pipeline, X)
-    x = assemble(X, env, target_pipe(p))
+    x = assemble(env, target_pipe(p), X)
     card = cardinality(target(x))
     optional = fits(x0to1, card)
     assemble_lift(p, optional ? minimum_missing : minimum, Pipeline[x])
 end
+
+translate(mod::Module, ::Val{:count}, args::Tuple{Any}) =
+    Count(translate(mod, args[1]))
+
+translate(mod::Module, ::Val{:count}, args::Tuple{}) =
+    Then(Count)
+
+translate(mod::Module, ::Val{:sum}, args::Tuple{Any}) =
+    Sum(translate(mod, args[1]))
+
+translate(mod::Module, ::Val{:sum}, args::Tuple{}) =
+    Then(Sum)
+
+translate(mod::Module, ::Val{:max}, args::Tuple{Any}) =
+    Max(translate(mod, args[1]))
+
+translate(mod::Module, ::Val{:max}, args::Tuple{}) =
+    Then(Max)
+
+translate(mod::Module, ::Val{:min}, args::Tuple{Any}) =
+    Min(translate(mod, args[1]))
+
+translate(mod::Module, ::Val{:min}, args::Tuple{}) =
+    Then(Min)
 
 
 #
@@ -1508,9 +1649,12 @@ Filter(X) =
     Query(Filter, X)
 
 function Filter(env::Environment, p::Pipeline, X)
-    x = assemble(X, env, target_pipe(p))
+    x = assemble(env, target_pipe(p), X)
     assemble_filter(p, x)
 end
+
+translate(mod::Module, ::Val{:filter}, args::Tuple{Any}) =
+    Filter(translate(mod, args[1]))
 
 
 #
@@ -1593,12 +1737,18 @@ Take(env::Environment, p::Pipeline, n::Union{Int,Missing}, inv::Bool=false) =
     assemble_take(p, n, inv)
 
 function Take(env::Environment, p::Pipeline, N, inv::Bool=false)
-    n = assemble(N, env, source_pipe(p))
+    n = assemble(env, source_pipe(p), N)
     assemble_take(p, n, inv)
 end
 
 Drop(env::Environment, p::Pipeline, N) =
     Take(env, p, N, true)
+
+translate(mod::Module, ::Val{:take}, args::Tuple{Any}) =
+    Take(translate(mod, args[1]))
+
+translate(mod::Module, ::Val{:drop}, args::Tuple{Any}) =
+    Drop(translate(mod, args[1]))
 
 
 #
@@ -1618,9 +1768,15 @@ Lift(::typeof(Unique)) =
     Then(Unique)
 
 function Unique(env::Environment, p::Pipeline, X)
-    x = assemble(X, env, target_pipe(p))
+    x = assemble(env, target_pipe(p), X)
     assemble_unique(p, x)
 end
+
+translate(mod::Module, ::Val{:unique}, args::Tuple{Any}) =
+    Unique(translate(mod, args[1]))
+
+translate(mod::Module, ::Val{:unique}, args::Tuple{}) =
+    Then(Unique)
 
 function assemble_group(p::Pipeline, xs::Vector{Pipeline})
     lbls = Symbol[]
@@ -1678,7 +1834,73 @@ Group(Xs...) =
     Query(Group, Xs...)
 
 function Group(env::Environment, p::Pipeline, Xs...)
-    xs = assemble.(collect(AbstractQuery, Xs), Ref(env), Ref(target_pipe(p)))
+    xs = assemble.(Ref(env), Ref(target_pipe(p)), collect(AbstractQuery, Xs))
     assemble_group(p, xs)
 end
+
+translate(mod::Module, ::Val{:group}, args::Tuple) =
+    Group(translate.(Ref(mod), args)...)
+
+
+#
+# Cardinality assertions.
+#
+
+function assemble_cardinality(p::Pipeline, card::Cardinality)
+    src = source(p)
+    tgt = BlockOf(elements(target(p)), card) |> IsFlow
+    chain_of(p, cardinality(card)) |> designate(src, tgt)
+end
+
+Is0to1(X) = Query(Is0to1, X)
+
+Is0toN(X) = Query(Is0toN, X)
+
+Is1to1(X) = Query(Is1to1, X)
+
+Is1toN(X) = Query(Is1toN, X)
+
+Lift(::typeof(Is0to1)) = Then(Is0to1)
+
+Lift(::typeof(Is0toN)) = Then(Is0toN)
+
+Lift(::typeof(Is1to1)) = Then(Is1to1)
+
+Lift(::typeof(Is1toN)) = Then(Is1toN)
+
+Is0to1(env::Environment, p::Pipeline, X) =
+    assemble_cardinality(assemble(X, env, p), x0to1)
+
+Is0toN(env::Environment, p::Pipeline, X) =
+    assemble_cardinality(assemble(X, env, p), x0toN)
+
+Is1to1(env::Environment, p::Pipeline, X) =
+    assemble_cardinality(assemble(X, env, p), x1to1)
+
+Is1toN(env::Environment, p::Pipeline, X) =
+    assemble_cardinality(assemble(X, env, p), x1toN)
+
+translate(mod::Module, ::Val{:is0to1}, args::Tuple{Any}) =
+    Is0to1(translate(mod, args[1]))
+
+translate(mod::Module, ::Val{:is0toN}, args::Tuple{Any}) =
+    Is0toN(translate(mod, args[1]))
+
+translate(mod::Module, ::Val{:is1to1}, args::Tuple{Any}) =
+    Is1to1(translate(mod, args[1]))
+
+translate(mod::Module, ::Val{:is1toN}, args::Tuple{Any}) =
+    Is1toN(translate(mod, args[1]))
+
+translate(mod::Module, ::Val{:is0to1}, args::Tuple{}) =
+    Then(Is0to1)
+
+translate(mod::Module, ::Val{:is0toN}, args::Tuple{}) =
+    Then(Is0toN)
+
+translate(mod::Module, ::Val{:is1to1}, args::Tuple{}) =
+    Then(Is1to1)
+
+translate(mod::Module, ::Val{:is1toN}, args::Tuple{}) =
+    Then(Is1toN)