+++ NEWS.md
@@ -1,21 +1,31 @@
 # Release Notes
 
+## v0.6.0
+
+- Added Cardinality assertion combinators.
+
+## v0.5.0
+
+- Added `@query` notation.
+- Fixed Tables.jl integration.
+
+## v0.4.0
+
+- Added `Collect` combinator.
+- Better rendering of oversized tables.
 
 ## v0.3.0
 
 - Import/export using Tables.jl interface.
 - Added `DataKnot(::Pair...)` constructor.
 
-
 ## v0.2.0
 
 - Added `Unique` and `Group` combinators.
 
-
 ## v0.1.1
 
 - Fixed packaging issues.
-
 
 ## v0.1.0